Fix core selector tests so that they pass in browsers with querySelectorAll.
When a multiple selector is passed to Ojay's default selector engine, YAHOO.util.Selector, it returns elements in the order the selectors were passed in, rather than document order. This means that a test for duplicate elements, which passes when using the YUI selector engine fails when using querySelectorAll. This commit fixes the test only; Ojay's selector functionality should not be assumed to return elements in any particular order.